move generatePrettyTable to its own file (no changes)
