+++ src/webgpu/util/check_contents.ts
@@ -8,6 +8,7 @@ import {
 } from '../../common/util/util.js';
 
 import { float16BitsToFloat32 } from './conversion.js';
+import { generatePrettyTable } from './pretty_diff_tables.js';
 
 /** Generate an expected value at `index`, to test for equality with the actual value. */
 export type CheckElementsGenerator = (index: number) => number;
@@ -235,54 +236,4 @@ function intToPaddedHex(number: number, { byteLength }: { byteLength: number }) 
   if (byteLength) s = s.padStart(byteLength * 2, '0');
   if (number < 0) s = '-' + s;
   return s;
-}
-
-/**
- * Pretty-prints a "table" of cell values (each being `number | string`), right-aligned.
- * Each row may be any iterator, including lazily-generated (potentially infinite) rows.
- *
- * The first argument is the printing options:
- *  - fillToWidth: Keep printing columns (as long as there is data) until this width is passed.
- *    If there is more data, "..." is appended.
- *  - numberToString: if a cell value is a number, this is used to stringify it.
- *
- * Each remaining argument provides one row for the table.
- */
-function generatePrettyTable(
-  { fillToWidth, numberToString }: { fillToWidth: number; numberToString: (n: number) => string },
-  rows: ReadonlyArray<Iterable<string | number>>
-): string {
-  const rowStrings = range(rows.length, () => '');
-  let totalTableWidth = 0;
-  const iters = rows.map(row => row[Symbol.iterator]());
-
-  // Loop over columns
-  for (;;) {
-    const cellsForColumn = iters.map(iter => {
-      const r = iter.next(); // Advance the iterator for each row, in lock-step.
-      return r.done ? undefined : typeof r.value === 'number' ? numberToString(r.value) : r.value;
-    });
-    if (cellsForColumn.every(cell => cell === undefined)) break;
-
-    // Maximum width of any cell in this column, plus one for space between columns
-    // (also inserts a space at the left of the first column).
-    const colWidth = Math.max(...cellsForColumn.map(c => (c === undefined ? 0 : c.length))) + 1;
-    for (let row = 0; row < rowStrings.length; ++row) {
-      const cell = cellsForColumn[row];
-      if (cell !== undefined) {
-        rowStrings[row] += cell.padStart(colWidth);
-      }
-    }
-
-    totalTableWidth += colWidth;
-    if (totalTableWidth >= fillToWidth) {
-      for (let row = 0; row < rowStrings.length; ++row) {
-        if (cellsForColumn[row] !== undefined) {
-          rowStrings[row] += ' ...';
-        }
-      }
-      break;
-    }
-  }
-  return rowStrings.join('\n');
 }
